update tractor trailer

+++ database/migrations/2024_07_18_074721_create_positions_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->string('description')->nullable();
             $table->tinyInteger('is_active')->nullable();
             $table->tinyInteger('is_deleted')->nullable();
-            $table->tinyInteger('deleted_by')->nullable();
+            $table->integer('deleted_by')->nullable();
             $table->tinyInteger('deleted_at')->nullable();
             $table->unsignedBigInteger('created_by')->nullable();
             $table->unsignedBigInteger('updated_by')->nullable();
